requirements.txt
django
djangorestframework
djongo
django-cors-headers
mongo